Extract out parse_ip.rs
